% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_device_meta}
\alias{get_device_meta}
\title{Get Device Metadata}
\usage{
get_device_meta(f)
}
\arguments{
\item{f}{path to fit file}
}
\description{
Get Device Metadata
}
\examples{
get_device_meta(system.file("fit/zwift/wahoo_h3.fit", package = "dualR"))
}
