% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_devices.R
\name{get_devices_summary}
\alias{get_devices_summary}
\title{get devices summary}
\usage{
get_devices_summary(devices, iv_mode = "")
}
\arguments{
\item{devices}{data.table of devices or path to fit file where devices can be retrieved.}

\item{iv_mode}{character "primary" or "secondary" for trainer power or power meter.}
}
\description{
given device table from a fit file, return a summary of connected devices.
}
