% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_fit_meta}
\alias{get_fit_meta}
\title{Get Fit Metadata}
\usage{
get_fit_meta(f)
}
\arguments{
\item{f}{path to fit file}
}
\description{
Get Fit Metadata
}
\examples{
get_fit_meta(system.file("fit/fit2.fit", package = "dualR"))
}
