% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_maxes}
\alias{get_maxes}
\title{get max average powers from fit power time series}
\usage{
get_maxes(fit, wthn = NULL)
}
\arguments{
\item{fit}{data.frame containing a "datetime" and "power" column}

\item{wthn}{second data.frame that overlaps the first.}
}
\description{
Creates a "critical power" curve from a timeseries of power data.
}
\examples{
get_maxes(read_fit_file(system.file("fit/fit1.fit", package = "dualR")))
get_maxes(read_fit_file(system.file("fit/fit1.fit", package = "dualR")),
          read_fit_file(system.file("fit/fit2.fit", package = "dualR")))
}
