% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_overlapping}
\alias{get_overlapping}
\title{get overlapping portion of timeseries}
\usage{
get_overlapping(fit, wthn)
}
\arguments{
\item{fit}{data.frame containing a "datetime" and "power" column}

\item{wthn}{second data.frame that overlaps the first.}
}
\description{
get overlapping portion of timeseries
}
