% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work.R
\name{get_work_summary}
\alias{get_work_summary}
\title{work summary}
\usage{
get_work_summary(
  fit,
  wthn = NULL,
  bins = c(seq(0, 1000, 50), 5000),
  labels = bins[1:length(bins) - 1]
)
}
\arguments{
\item{fit}{data.frame containing a "datetime" and "power" column}

\item{wthn}{second data.frame that overlaps the first.}

\item{bins}{numeric sequence defining bin edges}

\item{labels}{length(bins) - 1 length labels for output}
}
\value{
data.frame with bin (factor) and work (numeric) columns
}
\description{
get summary of kj generated from a power time series
}
\examples{
get_work_summary(read_fit_file(system.file("fit/fit1.fit", package = "dualR")))
get_work_summary(read_fit_file(system.file("fit/fit1.fit", package = "dualR")),
                 read_fit_file(system.file("fit/fit2.fit", package = "dualR")))
}
