% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{prep_fit}
\alias{prep_fit}
\title{prepare fit files for comparison}
\usage{
prep_fit(fit, wthn = NULL)
}
\arguments{
\item{fit}{data.frame containing a "datetime" and "power" column}

\item{wthn}{second data.frame that overlaps the first.}
}
\description{
prepare fit files for comparison
}
