% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{read_fit_file}
\alias{read_fit_file}
\title{Read Fit File}
\usage{
read_fit_file(f)
}
\arguments{
\item{f}{path to fit file}
}
\description{
Read Fit File
}
\examples{
read_fit_file(system.file("fit/fit1.fit", package = "dualR"))
}
