% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_devices.R
\name{get_devices_summary}
\alias{get_devices_summary}
\title{get devices summary}
\usage{
get_devices_summary(devices)
}
\arguments{
\item{devices}{data.table of devices or path to fit file where devices can be retrieved.}
}
\description{
given device table from a fit file, return a summary of connected devices.
}
