% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{get_fit_meta}
\alias{get_fit_meta}
\title{Get Fit Metadata}
\usage{
get_fit_meta(f)
}
\arguments{
\item{f}{path to fit file of FitFile object}
}
\description{
Get Fit Metadata
}
\examples{
f <- readFitFile(system.file("fit/fit1.fit", package = "dualR"))
get_fit_meta(f)

}
